Explicit casting of double->float in several places. Replaced NULL with nullptr. Replaced auto_ptr with unique_ptr

+++ source/TBReco/src/FastSimulation.cc
@@ -245,7 +245,7 @@ namespace depfet {
           double hitPos[3] =  {u , v , 0} ;
           simHit->setPosition(hitPos);  
           // Set local hit momentum 
-          float hitMom[3] = { dudw*mom/std::sqrt(dudw*dudw + dvdw*dvdw +1), dvdw*mom/std::sqrt(dudw*dudw + dvdw*dvdw +1) , 1.0*mom/std::sqrt(dudw*dudw + dvdw*dvdw +1) };
+          float hitMom[3] = { float(dudw*mom/std::sqrt(dudw*dudw + dvdw*dvdw +1)), float(dvdw*mom/std::sqrt(dudw*dudw + dvdw*dvdw +1)) , float(1.0*mom/std::sqrt(dudw*dudw + dvdw*dvdw +1))};
           simHit->setMomentum(hitMom);
           
           // Set CellID
